Updated to aop's mini
