+++ src/test/scala/mini/ALUTests.scala
@@ -3,11 +3,12 @@
 package mini
 
 import chisel3._
-import chisel3.util._
 import chisel3.testers._
-import ALU._
+import chisel3.util._
+import mini._
 
 class ALUTester(alu: => ALU)(implicit p: freechips.rocketchip.config.Parameters) extends BasicTester with TestUtils {
+  import ALU._
   val dut = Module(alu)
   val ctrl = Module(new Control)
   val xlen = p(XLEN)
