+++ src/test/scala/mini/BrCondTests.scala
@@ -3,11 +3,12 @@
 package mini
 
 import chisel3._
-import chisel3.util._
 import chisel3.testers._
-import Control._
+import chisel3.util._
+import mini._
 
 class BrCondTester(br: => BrCond)(implicit p: freechips.rocketchip.config.Parameters) extends BasicTester with TestUtils {
+  import Control._
   val dut = Module(br)
   val ctrl = Module(new Control)
   val xlen = p(XLEN)
